Add 在庫情報の更新機能を追加
在庫数のデータ型はbigdecimal型だと表記が不自然になるため、
flaotに統一した。

+++ src/main/java/com/example/demo/Form/InventoryForm.java
@@ -1,6 +1,5 @@
 package com.example.demo.Form;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import javax.validation.constraints.NotNull;
@@ -16,11 +15,11 @@ public class InventoryForm {
 
 
     @NotNull(message ="商品名を選択してください。")
-    private Integer productId;
+    private int productId;
 
     @NotNull(message = "数値を入力してください。")
     @PositiveOrZero(message = "正の数を入力してください。")
-    private BigDecimal quantity;
+    private float quantity;
 
     @NotNull(message = "日付を設定してください。")
     @DateTimeFormat(pattern = "yyyy-MM-dd")
